Clean code: Move CalendarMapper to calendar package.
